[ES|QL] separate `KEEP`, `DROP`, and `SORT` autocomplete routines (#197744)
## Summary

This PR begins the refactor described in
https://github.com/elastic/kibana/issues/195418.

The autocomplete engine now delegates to command-specific routines
attached to the command definitions for `KEEP`, `DROP`, and `SORT`.

The naming of `getFieldsFor` has been broadened to `getColumnsFor`
because the response from Elasticsearch can contain variables as well as
fields, depending on the query that is used to fetch the columns.

No user-facing behavior should have changed.

### Checklist

- [x] [Unit or functional
tests](https://www.elastic.co/guide/en/kibana/master/development-tests.html)
were updated or added to match the most common scenarios

---------

(cherry picked from commit 11ae6a5bd9a06a4402e8af5173b0b0efcf5f52fc)
