+++ packages/kbn-esql-validation-autocomplete/src/code_actions/actions.ts
@@ -403,7 +403,7 @@ export async function getActions(
   const { getPolicies, getPolicyFields } = getPolicyRetriever(resourceRetriever);
 
   const callbacks = {
-    getFieldsByType: resourceRetriever?.getFieldsFor ? getFieldsByType : undefined,
+    getFieldsByType: resourceRetriever?.getColumnsFor ? getFieldsByType : undefined,
     getSources: resourceRetriever?.getSources ? getSources : undefined,
     getPolicies: resourceRetriever?.getPolicies ? getPolicies : undefined,
     getPolicyFields: resourceRetriever?.getPolicies ? getPolicyFields : undefined,
