+++ packages/kbn-esql-validation-autocomplete/src/definitions/commands.ts
@@ -32,6 +32,9 @@ import {
   withOption,
 } from './options';
 import type { CommandDefinition } from './types';
+import { suggest as suggestForSort } from '../autocomplete/commands/sort';
+import { suggest as suggestForKeep } from '../autocomplete/commands/keep';
+import { suggest as suggestForDrop } from '../autocomplete/commands/drop';
 
 const statsValidator = (command: ESQLCommand) => {
   const messages: ESQLMessage[] = [];
@@ -148,7 +151,7 @@ const statsValidator = (command: ESQLCommand) => {
   }
   return messages;
 };
-export const commandDefinitions: CommandDefinition[] = [
+export const commandDefinitions: Array<CommandDefinition<any>> = [
   {
     name: 'row',
     description: i18n.translate('kbn-esql-validation-autocomplete.esql.definitions.rowDoc', {
@@ -311,6 +314,7 @@ export const commandDefinitions: CommandDefinition[] = [
       defaultMessage: 'Rearranges fields in the input table by applying the keep clauses in fields',
     }),
     examples: ['… | keep a', '… | keep a,b'],
+    suggest: suggestForKeep,
     options: [],
     modes: [],
     signature: {
@@ -330,6 +334,7 @@ export const commandDefinitions: CommandDefinition[] = [
       multipleParams: true,
       params: [{ name: 'column', type: 'column', wildcards: true }],
     },
+    suggest: suggestForDrop,
     validate: (command: ESQLCommand) => {
       const messages: ESQLMessage[] = [];
       const wildcardItems = command.args.filter((arg) => isColumnItem(arg) && arg.name === '*');
@@ -386,7 +391,9 @@ export const commandDefinitions: CommandDefinition[] = [
       multipleParams: true,
       params: [{ name: 'expression', type: 'any' }],
     },
+    suggest: suggestForSort,
   },
+
   {
     name: 'where',
     description: i18n.translate('kbn-esql-validation-autocomplete.esql.definitions.whereDoc', {
