+++ packages/kbn-esql-validation-autocomplete/src/definitions/types.ts
@@ -8,6 +8,7 @@
  */
 
 import type { ESQLCommand, ESQLCommandOption, ESQLFunction, ESQLMessage } from '@kbn/esql-ast';
+import { GetColumnsByTypeFn, SuggestionRawDefinition } from '../autocomplete/types';
 
 /**
  * All supported field types in ES|QL. This is all the types
@@ -158,14 +159,21 @@ export interface FunctionDefinition {
   validate?: (fnDef: ESQLFunction) => ESQLMessage[];
 }
 
-export interface CommandBaseDefinition {
-  name: string;
+export interface CommandBaseDefinition<CommandName extends string> {
+  name: CommandName;
   alias?: string;
   description: string;
   /**
    * Whether to show or hide in autocomplete suggestion list
    */
   hidden?: boolean;
+  suggest?: (
+    innerText: string,
+    command: ESQLCommand<CommandName>,
+    getColumnsByType: GetColumnsByTypeFn,
+    columnExists: (column: string) => boolean
+  ) => Promise<SuggestionRawDefinition[]>;
+  /** @deprecated this property will disappear in the future */
   signature: {
     multipleParams: boolean;
     // innerTypes here is useful to drill down the type in case of "column"
@@ -183,7 +191,8 @@ export interface CommandBaseDefinition {
   };
 }
 
-export interface CommandOptionsDefinition extends CommandBaseDefinition {
+export interface CommandOptionsDefinition<CommandName extends string = string>
+  extends CommandBaseDefinition<CommandName> {
   wrapped?: string[];
   optional: boolean;
   skipCommonValidation?: boolean;
@@ -201,12 +210,15 @@ export interface CommandModeDefinition {
   prefix?: string;
 }
 
-export interface CommandDefinition extends CommandBaseDefinition {
-  options: CommandOptionsDefinition[];
+export interface CommandDefinition<CommandName extends string>
+  extends CommandBaseDefinition<CommandName> {
   examples: string[];
   validate?: (option: ESQLCommand) => ESQLMessage[];
-  modes: CommandModeDefinition[];
   hasRecommendedQueries?: boolean;
+  /** @deprecated this property will disappear in the future */
+  modes: CommandModeDefinition[];
+  /** @deprecated this property will disappear in the future */
+  options: CommandOptionsDefinition[];
 }
 
 export interface Literals {
