+++ packages/kbn-esql-validation-autocomplete/src/shared/helpers.test.ts
@@ -9,7 +9,7 @@
 
 import { parse } from '@kbn/esql-ast';
 import { getExpressionType, shouldBeQuotedSource } from './helpers';
-import { SupportedDataType } from '../definitions/types';
+import type { SupportedDataType } from '../definitions/types';
 import { setTestFunctions } from './test_functions';
 
 describe('shouldBeQuotedSource', () => {
