+++ packages/kbn-esql-validation-autocomplete/src/validation/validation.test.ts
@@ -1532,7 +1532,7 @@ describe('validation logic', () => {
       it(`should not fetch source and fields list when a row command is set`, async () => {
         const callbackMocks = getCallbackMocks();
         await validateQuery(`row a = 1 | eval a`, getAstAndSyntaxErrors, undefined, callbackMocks);
-        expect(callbackMocks.getFieldsFor).not.toHaveBeenCalled();
+        expect(callbackMocks.getColumnsFor).not.toHaveBeenCalled();
         expect(callbackMocks.getSources).not.toHaveBeenCalled();
       });
 
@@ -1545,7 +1545,7 @@ describe('validation logic', () => {
       it(`should not fetch source and fields for empty command`, async () => {
         const callbackMocks = getCallbackMocks();
         await validateQuery(` `, getAstAndSyntaxErrors, undefined, callbackMocks);
-        expect(callbackMocks.getFieldsFor).not.toHaveBeenCalled();
+        expect(callbackMocks.getColumnsFor).not.toHaveBeenCalled();
         expect(callbackMocks.getSources).not.toHaveBeenCalled();
       });
 
@@ -1559,8 +1559,8 @@ describe('validation logic', () => {
         );
         expect(callbackMocks.getSources).not.toHaveBeenCalled();
         expect(callbackMocks.getPolicies).toHaveBeenCalled();
-        expect(callbackMocks.getFieldsFor).toHaveBeenCalledTimes(1);
-        expect(callbackMocks.getFieldsFor).toHaveBeenLastCalledWith({
+        expect(callbackMocks.getColumnsFor).toHaveBeenCalledTimes(1);
+        expect(callbackMocks.getColumnsFor).toHaveBeenLastCalledWith({
           query: `from enrich_index | keep otherField, yetAnotherField`,
         });
       });
@@ -1575,8 +1575,8 @@ describe('validation logic', () => {
         );
         expect(callbackMocks.getSources).not.toHaveBeenCalled();
         expect(callbackMocks.getPolicies).not.toHaveBeenCalled();
-        expect(callbackMocks.getFieldsFor).toHaveBeenCalledTimes(1);
-        expect(callbackMocks.getFieldsFor).toHaveBeenLastCalledWith({
+        expect(callbackMocks.getColumnsFor).toHaveBeenCalledTimes(1);
+        expect(callbackMocks.getColumnsFor).toHaveBeenLastCalledWith({
           query: 'show info',
         });
       });
@@ -1591,8 +1591,8 @@ describe('validation logic', () => {
         );
         expect(callbackMocks.getSources).toHaveBeenCalled();
         expect(callbackMocks.getPolicies).toHaveBeenCalled();
-        expect(callbackMocks.getFieldsFor).toHaveBeenCalledTimes(2);
-        expect(callbackMocks.getFieldsFor).toHaveBeenLastCalledWith({
+        expect(callbackMocks.getColumnsFor).toHaveBeenCalledTimes(2);
+        expect(callbackMocks.getColumnsFor).toHaveBeenLastCalledWith({
           query: `from enrich_index | keep otherField, yetAnotherField`,
         });
       });
@@ -1604,7 +1604,7 @@ describe('validation logic', () => {
             getAstAndSyntaxErrors,
             undefined,
             {
-              getFieldsFor: undefined,
+              getColumnsFor: undefined,
               getSources: undefined,
               getPolicies: undefined,
             }
@@ -1718,7 +1718,7 @@ describe('validation logic', () => {
       const contentByCallback = {
         getSources: /Unknown index/,
         getPolicies: /Unknown policy/,
-        getFieldsFor: /Unknown column|Argument of|it is unsupported or not indexed/,
+        getColumnsFor: /Unknown column|Argument of|it is unsupported or not indexed/,
         getPreferences: /Unknown/,
         getFieldsMetadata: /Unknown/,
       };
@@ -1761,7 +1761,7 @@ describe('validation logic', () => {
     });
 
     // test excluding one callback at the time
-    it.each(['getSources', 'getFieldsFor', 'getPolicies'] as Array<keyof typeof ignoreErrorsMap>)(
+    it.each(['getSources', 'getColumnsFor', 'getPolicies'] as Array<keyof typeof ignoreErrorsMap>)(
       `should not error if %s is missing`,
       async (excludedCallback) => {
         const filteredTestCases = fixtures.testCases.filter((t) =>
@@ -1790,7 +1790,7 @@ describe('validation logic', () => {
     );
 
     it('should work if no callback passed', async () => {
-      const excludedCallbacks = ['getSources', 'getPolicies', 'getFieldsFor'] as Array<
+      const excludedCallbacks = ['getSources', 'getPolicies', 'getColumnsFor'] as Array<
         keyof typeof ignoreErrorsMap
       >;
       for (const testCase of fixtures.testCases.filter((t) =>
